refactor(exposure): drop thin wrappers, call AsyncRecord directly
Collapse pipeline_record, flag_resolve, exposure_validate, and
exposure_context into exposure.go. Eval and exposure both use
GetDataRecorder().AsyncRecord when recorder gates pass; Datar/Kafka
check recordSource inline.
