+++ pkg/handler/data_recorder_datar.go
@@ -19,8 +19,7 @@ type datarRecorder struct {
 }
 
 func (d *datarRecorder) AsyncRecord(r models.EvalResult) {
-	// Datar aggregates evaluation assignments only; exposure rows use the same EvalResult wire shape.
-	if isExposurePipelineEvent(r) {
+	if r.RecordSource == models.EvalResultRecordSourceExposure {
 		return
 	}
 	d.engine.Record(r.FlagID, r.VariantID, r.SegmentID)
