+++ pkg/handler/data_recorder_kafka.go
@@ -160,7 +160,7 @@ func (k *kafkaRecorder) AsyncRecord(r models.EvalResult) {
 }
 
 var logKafkaAsyncRecordToDatadog = func(r models.EvalResult) {
-	if config.Global.StatsdClient == nil || isExposurePipelineEvent(r) {
+	if config.Global.StatsdClient == nil || r.RecordSource == models.EvalResultRecordSourceExposure {
 		return
 	}
 	config.Global.StatsdClient.Incr(
