+++ pkg/handler/exposure_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/openflagr/flagr/pkg/config"
+	"github.com/openflagr/flagr/pkg/entity"
 	"github.com/openflagr/flagr/swagger_gen/models"
 	exposureapi "github.com/openflagr/flagr/swagger_gen/restapi/operations/exposure"
 	"github.com/prashantv/gostub"
@@ -72,11 +73,10 @@ func TestPostExposures_RecordsWhenEnabled(t *testing.T) {
 	defer gostub.StubFunc(&GetEvalCache, cache).Reset()
 	defer gostub.Stub(&config.Config.RecorderEnabled, true).Reset()
 
-	var recorded []models.EvalResult
-	defer gostub.Stub(&recordPipelineEvent, func(r models.EvalResult) {
-		recorded = append(recorded, r)
-	}).Reset()
-
+	mock := &mockRecorder{}
+	prev := singletonDataRecorder
+	singletonDataRecorder = fanOutRecorder{mock}
+	defer func() { singletonDataRecorder = prev }()
 	h := NewExposure()
 	eid := "user-1"
 	body := &models.ExposuresRequest{
@@ -90,15 +90,60 @@ func TestPostExposures_RecordsWhenEnabled(t *testing.T) {
 		return
 	}
 	assert.Equal(t, int64(1), okRes.Payload.LoggedCount)
-	if assert.Len(t, recorded, 1) {
-		assert.Equal(t, models.EvalResultRecordSourceExposure, recorded[0].RecordSource)
+	if assert.Len(t, mock.calls, 1) {
+		assert.Equal(t, models.EvalResultRecordSourceExposure, mock.calls[0].RecordSource)
 	}
 }
 
-func TestProcessExposureRow_NullRow(t *testing.T) {
-	n, err := processExposureRow(0, nil)
-	assert.Equal(t, int64(0), n)
-	assert.NotNil(t, err)
+func TestValidateAndBuildExposure_FlagResolution(t *testing.T) {
+	defer gostub.StubFunc(&GetEvalCache, GenFixtureEvalCache()).Reset()
+	fixture := entity.GenFixtureFlag()
+	eid := "e1"
+
+	t.Run("by flag id", func(t *testing.T) {
+		r, err := validateAndBuildExposure(&models.Exposure{EntityID: &eid, FlagID: int64(fixture.ID)})
+		if !assert.NoError(t, err) {
+			return
+		}
+		assert.Equal(t, int64(fixture.ID), r.FlagID)
+	})
+
+	t.Run("missing flag ref", func(t *testing.T) {
+		_, err := validateAndBuildExposure(&models.Exposure{EntityID: &eid})
+		assert.Error(t, err)
+	})
+
+	t.Run("mismatched id and key", func(t *testing.T) {
+		f1 := entity.GenFixtureFlag()
+		f2 := entity.GenFixtureFlag()
+		f2.ID = 101
+		f2.Key = "flag_key_101"
+		ec := &EvalCache{cache: &cacheContainer{
+			idCache:  map[string]*entity.Flag{"100": &f1, "101": &f2},
+			keyCache: map[string]*entity.Flag{f1.Key: &f1, f2.Key: &f2},
+		}}
+		defer gostub.StubFunc(&GetEvalCache, ec).Reset()
+		_, err := validateAndBuildExposure(&models.Exposure{EntityID: &eid, FlagID: int64(f1.ID), FlagKey: f2.Key})
+		assert.Error(t, err)
+	})
+}
+
+func TestExposureVariant(t *testing.T) {
+	flag := entity.GenFixtureFlag()
+	id, key, err := exposureVariant(&flag, 300, "")
+	if !assert.NoError(t, err) {
+		return
+	}
+	assert.Equal(t, int64(300), id)
+	assert.Equal(t, "control", key)
+}
+
+func TestExposureMergeJSONMap(t *testing.T) {
+	dst := map[string]interface{}{}
+	exposureMergeJSONMap(dst, map[string]any{"a": 1})
+	exposureMergeJSONMap(dst, map[string]interface{}{"b": 2})
+	assert.Equal(t, float64(1), dst["a"])
+	assert.Equal(t, float64(2), dst["b"])
 }
 
 func strPtr(s string) *string { return &s }
